Remove SystemObjectModelProvider
